Make Windows path work on fake Windows
Added a hack to override Puppet's OS detection, based on the `kernel`
fact, as set in the test. Normally Puppet assumes (in some places where
it doesn't have access to the target OS's facts) that it is targetting
the OS that the specs are running on, not the OS specified by the facts
in the spec. Should work around rodjek/rspec-puppet#192 , though there
are more problems like this when using the `compile` matchers in specs
targetting Windows (e.g. packages want the `msi` provider to be
available).